Move sshkey to awsmodel
